requirements.txt
streamlit
groq
python-dotenv
pdfminer.six
sentence-transformers
scikit-learn
